Drop the Mega compat shim in favor of Torch calls and the cute backend utils
current_device/tensor_device_index/data_ptr/multiprocessor_count were one-line
Torch wrappers; get_device_properties duplicated attn_gym._backends.cute.utils.
validate_tma_tensor and initialized_cuda_device move to that module, and the
checkpoint capacity bound moves next to the other Mega host helpers.

stack-info: PR: https://github.com/meta-pytorch/attention-gym/pull/526, branch: drisspg/stack/156
