+++ attn_gym/linear/_delta_rule/mega/backward.py
@@ -7,8 +7,8 @@ from __future__ import annotations
 import torch
 
 from attn_gym._backends.cute import tensor_supports_contiguous_dim, tensor_supports_tma
+from attn_gym._backends.cute.utils import initialized_cuda_device
 from attn_gym.linear._delta_rule.mega.kernels.common.host import tensormap_workspace_bytes
-from attn_gym.linear._delta_rule.mega.kernels.compat import initialized_cuda_device
 from attn_gym.linear._delta_rule.validation import resolve_scale
 from attn_gym.utils import ceildiv
