+++ attn_gym/linear/_delta_rule/mega/kernels/gdn_bprop_f16.py
@@ -110,16 +110,17 @@ from functools import cache, partial
 from typing import NamedTuple
 
 import cuda.bindings.driver as cuda_driver
+import torch
 import cutlass
 import cutlass.experimental.primitives as nvvm
 from cutlass import cute
 from cutlass.experimental import cuda
 
 from attn_gym._backends.cute import compile_tvm_ffi
-from attn_gym._backends.cute.utils import requires_int64_abi
+from attn_gym._backends.cute.utils import get_device_properties, requires_int64_abi, validate_tma_tensor
 
 from .common.elementwise import softplus
-from .common.host import get_dtype
+from .common.host import checkpoint_capacity_bound, get_dtype
 from .common.split_k import (
     ORDER_CAPACITY,
     ORDER_ELEMS,
@@ -141,13 +142,7 @@ from .common.tvm_ffi import (
     make_workspace_signature,
     validate_cu_seqlens,
 )
-from .compat import (
-    checkpoint_capacity_bound,
-    current_device,
-    get_device_properties,
-    tensor_device_index,
-    validate_tma_tensor,
-)
+
 from .gdn_bprop_config import CFG
 from .tile_dsl.barrier import (
     MBarrier,
@@ -5609,8 +5604,8 @@ def chunk_gdn_bwd_sm100(
         sched_all,
         tensormap_workspace,
     )
-    device_index = tensor_device_index(q)
-    if current_device() != device_index:
+    device_index = q.get_device()
+    if torch.cuda.current_device() != device_index:
         raise ValueError("the active CUDA device must match q.device")
     device_properties = get_device_properties(device_index)
     num_sm = device_properties.multi_processor_count
