+++ attn_gym/linear/_delta_rule/mega/kernels/kda_bprop_f16.py
@@ -76,22 +76,17 @@ from functools import lru_cache, partial
 from typing import NamedTuple, Type
 
 import cuda.bindings.driver as cuda_driver
+import torch
 import cutlass
 import cutlass.experimental.cuda as cuda
 import cutlass.experimental.primitives as nvvm
 import cutlass.cute as cute
 
 from attn_gym._backends.cute import compile_tvm_ffi
-from attn_gym._backends.cute.utils import requires_int64_abi
+from attn_gym._backends.cute.utils import get_device_properties, requires_int64_abi, validate_tma_tensor
 from attn_gym.linear._delta_rule.mega.kernels.common.split_k import ORDER_CAPACITY, ORDER_ELEMS, ORDER_THREADS, decode_work_item, order_body
-from attn_gym.linear._delta_rule.mega.kernels.common.host import get_dtype
-from attn_gym.linear._delta_rule.mega.kernels.compat import (
-    checkpoint_capacity_bound,
-    current_device,
-    get_device_properties,
-    tensor_device_index,
-    validate_tma_tensor,
-)
+from attn_gym.linear._delta_rule.mega.kernels.common.host import checkpoint_capacity_bound, get_dtype
+
 from attn_gym.linear._delta_rule.mega.kernels.common.thd import (
     TENSOR_MAP_QWORDS,
     emit_checkpoint_seq_descs,
@@ -4324,8 +4319,8 @@ def chunk_kda_bwd_sm100(
         sched_all,
         tensormap_workspace,
     )
-    device_index = tensor_device_index(q)
-    if current_device() != device_index:
+    device_index = q.get_device()
+    if torch.cuda.current_device() != device_index:
         raise ValueError("the active CUDA device must match q.device")
     device_properties = get_device_properties(device_index)
     num_sm = device_properties.multi_processor_count
